[FIVE-301-XmlValidation] GP: Se movió el método ValidateSettings del servicio ArtifactsService a una nueva clase llamada SettingsValidator
